Adding product stats endpoint
    /productStats/:channel/:id

returns something that looks like this

{
  "TotalRevenue": 0,
  "TotalOrders": 0,
  "AverageNumberPerOrder": 0,
  "TotalInCarts": 18,
  "ProductConversionRate": 0,
  "AverageTotalRevenue": 0,
  "AverageTotalOrders": 0,
  "AverageAverageNumberPerOrder": 0,
  "AverageTotalInCarts": 9,
  "AverageProductConversionRate": 0
}

+++ intelligence/eggcrate/src/services/productFunnel.go
@@ -63,7 +63,7 @@ func buildResponse(pf responses.HenhouseResponse) string {
 		checkoutPurchased = float64(purchases) / float64(checkoutClicks)
 	}
 	if checkoutPurchased = 0.0; checkoutClicks > 0.0 {
-		checkoutPurchased = float32(purchases) / float32(checkoutClicks)
+		checkoutPurchased = float64(purchases) / float64(checkoutClicks)
 	}
 
 	resp := responses.ProductFunnelResponse{
